[dagster-iceberg] Centralize using one I/O manager
