--- libraries/dagster-iceberg/src/dagster_iceberg/io_manager/__init__.py
@@ -1,1 +1,0 @@
-from dagster_iceberg.io_manager.base import IcebergIOManager as IcebergIOManager
