+++ libraries/dagster-iceberg/tests/io_managers/test_daft_io_manager.py
@@ -15,14 +15,14 @@ from dagster import (
 from pyiceberg.catalog import Catalog
 
 from dagster_iceberg.config import IcebergCatalogConfig
-from dagster_iceberg.io_manager.daft import IcebergDaftIOManager
+from dagster_iceberg.io_manager import IcebergIOManager
 
 
 @pytest.fixture
 def io_manager(
     catalog_name: str, namespace: str, catalog_config_properties: Dict[str, str]
-) -> IcebergDaftIOManager:
-    return IcebergDaftIOManager(
+) -> IcebergIOManager:
+    return IcebergIOManager(
         name=catalog_name,
         config=IcebergCatalogConfig(properties=catalog_config_properties),
         namespace=namespace,
@@ -30,7 +30,7 @@ def io_manager(
 
 
 @pytest.fixture
-def custom_db_io_manager(io_manager: IcebergDaftIOManager):
+def custom_db_io_manager(io_manager: IcebergIOManager):
     return io_manager.create_io_manager(None)
 
 
@@ -133,7 +133,7 @@ def test_iceberg_io_manager_with_assets(
     asset_b_df_table_identifier: str,
     asset_b_plus_one_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergDaftIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -153,7 +153,7 @@ def test_iceberg_io_manager_with_assets(
 def test_iceberg_io_manager_with_daily_partitioned_assets(
     asset_daily_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergDaftIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -183,7 +183,7 @@ def test_iceberg_io_manager_with_daily_partitioned_assets(
 def test_iceberg_io_manager_with_hourly_partitioned_assets(
     asset_hourly_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergDaftIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -213,7 +213,7 @@ def test_iceberg_io_manager_with_hourly_partitioned_assets(
 def test_iceberg_io_manager_with_multipartitioned_assets(
     asset_multi_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergDaftIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
