+++ libraries/dagster-iceberg/tests/io_managers/test_pandas_io_manager.py
@@ -16,14 +16,14 @@ from dagster import (
 from pyiceberg.catalog import Catalog
 
 from dagster_iceberg.config import IcebergCatalogConfig
-from dagster_iceberg.io_manager.pandas import IcebergPandasIOManager
+from dagster_iceberg.io_manager import IcebergIOManager
 
 
 @pytest.fixture
 def io_manager(
     catalog_name: str, namespace: str, catalog_config_properties: Dict[str, str]
-) -> IcebergPandasIOManager:
-    return IcebergPandasIOManager(
+) -> IcebergIOManager:
+    return IcebergIOManager(
         name=catalog_name,
         config=IcebergCatalogConfig(properties=catalog_config_properties),
         namespace=namespace,
@@ -31,7 +31,7 @@ def io_manager(
 
 
 @pytest.fixture
-def custom_db_io_manager(io_manager: IcebergPandasIOManager):
+def custom_db_io_manager(io_manager: IcebergIOManager):
     return io_manager.create_io_manager(None)
 
 
@@ -139,7 +139,7 @@ def test_iceberg_io_manager_with_assets(
     asset_b_df_table_identifier: str,
     asset_b_plus_one_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPandasIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -159,7 +159,7 @@ def test_iceberg_io_manager_with_assets(
 def test_iceberg_io_manager_with_daily_partitioned_assets(
     asset_daily_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPandasIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -189,7 +189,7 @@ def test_iceberg_io_manager_with_daily_partitioned_assets(
 def test_iceberg_io_manager_with_hourly_partitioned_assets(
     asset_hourly_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPandasIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -219,7 +219,7 @@ def test_iceberg_io_manager_with_hourly_partitioned_assets(
 def test_iceberg_io_manager_with_multipartitioned_assets(
     asset_multi_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPandasIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
