+++ libraries/dagster-iceberg/tests/io_managers/test_polars_io_manager.py
@@ -15,14 +15,14 @@ from dagster import (
 from pyiceberg.catalog import Catalog
 
 from dagster_iceberg.config import IcebergCatalogConfig
-from dagster_iceberg.io_manager.polars import IcebergPolarsIOManager
+from dagster_iceberg.io_manager import IcebergIOManager
 
 
 @pytest.fixture
 def io_manager(
     catalog_name: str, namespace: str, catalog_config_properties: Dict[str, str]
-) -> IcebergPolarsIOManager:
-    return IcebergPolarsIOManager(
+) -> IcebergIOManager:
+    return IcebergIOManager(
         name=catalog_name,
         config=IcebergCatalogConfig(properties=catalog_config_properties),
         namespace=namespace,
@@ -30,7 +30,7 @@ def io_manager(
 
 
 @pytest.fixture
-def custom_db_io_manager(io_manager: IcebergPolarsIOManager):
+def custom_db_io_manager(io_manager: IcebergIOManager):
     return io_manager.create_io_manager(None)
 
 
@@ -137,7 +137,7 @@ def test_iceberg_io_manager_with_assets(
     asset_b_df_table_identifier: str,
     asset_b_plus_one_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPolarsIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -157,7 +157,7 @@ def test_iceberg_io_manager_with_assets(
 def test_iceberg_io_manager_with_daily_partitioned_assets(
     asset_daily_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPolarsIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -187,7 +187,7 @@ def test_iceberg_io_manager_with_daily_partitioned_assets(
 def test_iceberg_io_manager_with_hourly_partitioned_assets(
     asset_hourly_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPolarsIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
 
@@ -217,7 +217,7 @@ def test_iceberg_io_manager_with_hourly_partitioned_assets(
 def test_iceberg_io_manager_with_multipartitioned_assets(
     asset_multi_partitioned_table_identifier: str,
     catalog: Catalog,
-    io_manager: IcebergPolarsIOManager,
+    io_manager: IcebergIOManager,
 ):
     resource_defs = {"io_manager": io_manager}
